+++ source/ChanSort/GlobalImageCollection.cs
@@ -58,6 +58,7 @@ namespace ChanSort.Ui
       this.sharedImageCollection1.ImageSource.Images.SetKeyName(38, "0038.png");
       this.sharedImageCollection1.ImageSource.Images.SetKeyName(39, "0039.png");
       this.sharedImageCollection1.ImageSource.Images.SetKeyName(40, "0040.png");
+      this.sharedImageCollection1.ImageSource.Images.SetKeyName(41, "0041.png");
       this.sharedImageCollection1.ParentControl = null;
       ((System.ComponentModel.ISupportInitialize)(this.sharedImageCollection1.ImageSource)).EndInit();
       ((System.ComponentModel.ISupportInitialize)(this.sharedImageCollection1)).EndInit();